refactor: util 클래스 분리
