+++ tomcat/src/main/java/org/apache/coyote/http11/controller/AbstractController.java
@@ -1,8 +1,8 @@
 package org.apache.coyote.http11.controller;
 
-import static org.apache.coyote.http11.HttpUtils.getContentType;
-import static org.apache.coyote.http11.HttpUtils.readContentsFromFile;
+import static org.apache.coyote.http11.ParseUtils.parseContentType;
 import static org.apache.coyote.http11.header.HeaderType.ACCEPT;
+import static org.apache.coyote.http11.utils.IOUtils.readContentsFromFile;
 
 import java.io.IOException;
 import org.apache.coyote.http11.header.HeaderType;
@@ -14,6 +14,22 @@ import org.apache.coyote.http11.responseline.HttpStatus;
 import org.apache.coyote.http11.responseline.ResponseLine;
 
 public abstract class AbstractController implements Controller {
+
+  protected static HttpResponse responseStaticFile(final HttpRequest request, final String path)
+      throws IOException {
+    final String body = readContentsFromFile(path);
+    final String contentType = parseContentType(request.getHeader(ACCEPT));
+    final ResponseLine responseLine = new ResponseLine(HttpStatus.OK);
+    final HttpHeader header = new HttpHeader();
+    header.setHeader(HeaderType.CONTENT_TYPE, contentType + ";charset=utf-8");
+    header.setHeader(HeaderType.CONTENT_LENGTH, body.getBytes().length + "");
+    return new HttpResponse(responseLine, header, body);
+  }
+
+  protected static HttpResponse responseFoundRedirect(final HttpHeader header) {
+    final ResponseLine responseLine = new ResponseLine(HttpStatus.FOUND);
+    return new HttpResponse(responseLine, header);
+  }
 
   @Override
   public HttpResponse service(final HttpRequest request) throws IOException {
@@ -32,21 +48,5 @@ public abstract class AbstractController implements Controller {
 
   protected HttpResponse doGet(final HttpRequest request) throws IOException {
     throw new UnsupportedOperationException();
-  }
-
-  protected HttpResponse responseStaticFile(final HttpRequest request, final String path)
-      throws IOException {
-    final String body = readContentsFromFile(path);
-    final String contentType = getContentType(request.getHeader(ACCEPT));
-    final ResponseLine responseLine = new ResponseLine(HttpStatus.OK);
-    final HttpHeader header = new HttpHeader();
-    header.setHeader(HeaderType.CONTENT_TYPE, contentType + ";charset=utf-8");
-    header.setHeader(HeaderType.CONTENT_LENGTH, body.getBytes().length + "");
-    return new HttpResponse(responseLine, header, body);
-  }
-
-  protected HttpResponse responseFoundRedirect(final HttpHeader header) {
-    final ResponseLine responseLine = new ResponseLine(HttpStatus.FOUND);
-    return new HttpResponse(responseLine, header);
   }
 }
